Remove azure(canvas) dependency from canvas_traits
