+++ components/canvas_traits/Cargo.toml
@@ -10,7 +10,6 @@ name = "canvas_traits"
 path = "lib.rs"
 
 [dependencies]
-azure = {git = "https://github.com/servo/rust-azure", features = ["plugins"]}
 cssparser = {version = "0.7", features = ["heap_size", "serde-serialization"]}
 euclid = "0.10.1"
 heapsize = "0.3.0"
